restructuring contd
