+++ pkg/client/security.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package main
+package client
 
 import (
 	"net/http"
@@ -21,12 +21,12 @@ type CORSConfig struct {
 func LoadCORSConfigFromEnv() CORSConfig {
 	originsStr := os.Getenv("MCP_ALLOWED_ORIGINS")
 	mode := os.Getenv("MCP_CORS_MODE")
-	
+
 	// Default to strict mode if not specified
 	if mode == "" {
 		mode = "strict"
 	}
-	
+
 	var origins []string
 	if originsStr != "" {
 		origins = strings.Split(originsStr, ",")
@@ -35,7 +35,7 @@ func LoadCORSConfigFromEnv() CORSConfig {
 			origins[i] = strings.TrimSpace(origins[i])
 		}
 	}
-	
+
 	return CORSConfig{
 		AllowedOrigins: origins,
 		Mode:           mode,
@@ -48,26 +48,26 @@ func isOriginAllowed(origin string, allowedOrigins []string, mode string) bool {
 	if mode == "disabled" {
 		return true
 	}
-	
+
 	// Check if origin is in the allowed list
 	for _, allowed := range allowedOrigins {
 		if origin == allowed {
 			return true
 		}
 	}
-	
+
 	// In development mode, also allow localhost origins
 	if mode == "development" {
-		if strings.HasPrefix(origin, "http://localhost:") || 
-		   strings.HasPrefix(origin, "https://localhost:") ||
-		   strings.HasPrefix(origin, "http://127.0.0.1:") ||
-		   strings.HasPrefix(origin, "https://127.0.0.1:") ||
-		   strings.HasPrefix(origin, "http://[::1]:") ||
-		   strings.HasPrefix(origin, "https://[::1]:") {
+		if strings.HasPrefix(origin, "http://localhost:") ||
+			strings.HasPrefix(origin, "https://localhost:") ||
+			strings.HasPrefix(origin, "http://127.0.0.1:") ||
+			strings.HasPrefix(origin, "https://127.0.0.1:") ||
+			strings.HasPrefix(origin, "http://[::1]:") ||
+			strings.HasPrefix(origin, "https://[::1]:") {
 			return true
 		}
 	}
-	
+
 	return false
 }
 
@@ -89,24 +89,24 @@ func (h *securityHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 			http.Error(w, "Origin not allowed", http.StatusForbidden)
 			return
 		}
-		
+
 		// Log allowed origins at debug level to avoid too much noise in production
 		h.logger.Debugf("Allowed request from origin: %s", origin)
-		
+
 		// If we have a valid origin, add CORS headers
 		w.Header().Set("Access-Control-Max-Age", "3600")
 		w.Header().Set("Access-Control-Allow-Origin", origin)
 		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, OPTIONS")
 		w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Mcp-Session-Id")
 	}
-	
+
 	// Handle OPTIONS requests for CORS preflight
 	if r.Method == http.MethodOptions {
 		h.logger.Debugf("Handling OPTIONS preflight request from origin: %s", origin)
 		w.WriteHeader(http.StatusOK)
 		return
 	}
-	
+
 	// If origin is valid or not present, delegate to the wrapped handler
 	h.handler.ServeHTTP(w, r)
 }
